Add data source for payment method
Allow a payment method to be searched for, so that it can be used by the subscription resource.

Payment methods can either be filtered by card type or last numbers in the account number, or just use the only payment method available.

Also add credentials to allow tests to run during CI.