Add validation for length of database name
Right now terraform provider does not have any restriction on length of the database name.
But the redislabs api limits length of the database name to max of 40 characters. This causes `terraform apply` to fail.
This commit adds validation for the same.

Error observed (only when `TF_LOG=trace` is set):
```
"additionalInfo" : "Database name is longer than 40 characters"
```