fix: restructure regions resource to always use update logic, fix issue
Also some minor improvements such as code simplifications, and avoiding
the use of the `cmp` package which should only be used in tests as it
can panic instead of error for some conditions.

The create function now calls into the update function so that the logic
is consistent. It also rearranges a lot of the code so it is easier to
follow.

Finally the order of when things get created and deleted are more clear
and the appropriate waits are in place.

There is an outstanding TODO regarding subscription mutexes which I will
seek further clarification on before merging.