Add port, remote backup & CIDRs support (#371)
* Add port, remote backup & CIDRs support

Add support for:
* Specifying the port number on a normal or active/active database
* Configuring backups for normal or active/active databases
* Specifying multiple CIDR ranges when peering a normal or active/active
database

* Run all acceptance tests in parallel

* Stream `go test` logs

Go will only stream test logs if it is testing a single module at a
time, otherwise it will buffer the logs to the end and then display
them. This causes issues for CI as it means there's no idea what is
happening at the moment and the logs will be missing if the run is
cancelled.

See https://github.com/golang/go/issues/27826.

* Run all acceptance tests in parallel

CI is taking far too long and needs to be sped up.