Remove latest_backup_status and latest_import_status from resources
The `latest_backup_status` and `latest_import_status` attributes are
slow to compute and unnecessary for most users. Those attributes are
also not a great fit for a Terraform resource. Therefore, this PR
removes them to make the Plan/Apply much faster for the general case.

Users can still obtain those attributes via data sources.
The migration guide, also included in this PR, describes how to
migrate to them.